+++ CMake/SampleFileList.cmake
@@ -102,13 +102,9 @@ set(transform_SRC_FILES
 )
 
 set(lottie_HDR_FILES
-    ${PROJECT_SOURCE_DIR}/Samples/basic/lottie/src/ElementLottie.h
-    ${PROJECT_SOURCE_DIR}/Samples/basic/lottie/src/LottiePlugin.h
 )
 
 set(lottie_SRC_FILES
-    ${PROJECT_SOURCE_DIR}/Samples/basic/lottie/src/ElementLottie.cpp
-    ${PROJECT_SOURCE_DIR}/Samples/basic/lottie/src/LottiePlugin.cpp
     ${PROJECT_SOURCE_DIR}/Samples/basic/lottie/src/main.cpp
 )