+++ Source/Lottie/LottiePlugin.cpp
@@ -26,16 +26,48 @@
  *
  */
 
-#ifndef RMLUI_LOTTIE_PLUGIN_H
-#define RMLUI_LOTTIE_PLUGIN_H
 
+#include "../../Include/RmlUi/Lottie/ElementLottie.h"
+#include "../../Include/RmlUi/Core/Core.h"
+#include "../../Include/RmlUi/Core/ElementInstancer.h"
+#include "../../Include/RmlUi/Core/Factory.h"
+#include "../../Include/RmlUi/Core/Log.h"
+#include "../../Include/RmlUi/Core/Plugin.h"
 
 namespace Rml {
 namespace Lottie {
 
-void Initialise();
 
-}
+class LottiePlugin : public Plugin {
+public:
+	void OnInitialise() override
+	{
+		instancer = MakeUnique<ElementInstancerGeneric<ElementLottie> >();
+		
+		Factory::RegisterElementInstancer("lottie", instancer.get());
+
+		Log::Message(Log::LT_INFO, "Lottie plugin initialised.");
+	}
+
+	void OnShutdown() override
+	{
+		delete this;
+	}
+
+	int GetEventClasses() override
+	{
+		return Plugin::EVT_BASIC;
+	}
+
+private:
+	UniquePtr<ElementInstancerGeneric<ElementLottie>> instancer;
+};
+
+
+void Initialise()
+{
+	RegisterPlugin(new LottiePlugin);
 }
 
-#endif
+} // namespace Lottie
+} // namespace Rml